+++ src/lib/index.ts
@@ -2,5 +2,5 @@ export * from './ast/index.ts';
 export * from './renderer/index.ts';
 export * from './parser/index.ts';
 export * from './validation/index.ts';
-
+export {convertColor} from "./syntax/color/color.ts";
 export {isOkLabClose, okLabDistance} from "./syntax/color/utils/distance.ts";