+++ src/lib/validation/syntaxes/keyframe-selector.ts
@@ -7,32 +7,21 @@ export function validateKeyframeSelector(tokens: Token[], options: ValidationOpt
 
     consumeWhitespace(tokens);
 
-    if (tokens.length == 0) {
-
-        // @ts-ignore
-        return {
-            valid: SyntaxValidationResult.Drop,
-            context: [],
-            node: null,
-            syntax: null,
-            error: 'expected keyframe selector'
-        }
-    }
+    // if (tokens.length == 0) {
+    //
+    //     // @ts-ignore
+    //     return {
+    //         valid: SyntaxValidationResult.Drop,
+    //         context: [],
+    //         node: null,
+    //         syntax: null,
+    //         error: 'expected keyframe selector'
+    //     }
+    // }
 
     for (const t of splitTokenList(tokens)) {
 
-        if (t.length != 1) {
-
-            return {
-                valid: SyntaxValidationResult.Drop,
-                context: [],
-                node: t[0] ?? null,
-                syntax: null,
-                error: 'unexpected token'
-            }
-        }
-
-        if (t[0].typ != EnumToken.PercentageTokenType && !(t[0].typ == EnumToken.IdenTokenType && ['from', 'to', 'cover', 'contain', 'entry', 'exit', 'entry-crossing', 'exit-crossing'].includes((t[0] as IdentToken).val))) {
+        if (t.length != 1 || (t[0].typ != EnumToken.PercentageTokenType && !(t[0].typ == EnumToken.IdenTokenType && ['from', 'to', 'cover', 'contain', 'entry', 'exit', 'entry-crossing', 'exit-crossing'].includes((t[0] as IdentToken).val)))) {
 
             return {
                 valid: SyntaxValidationResult.Drop,