+++ src/lib/validation/syntaxes/url.ts
@@ -6,50 +6,49 @@ import {consumeWhitespace} from "../utils/index.ts";
 
 export function validateURL(token: Token): ValidationSyntaxResult {
 
-    if (token.typ == EnumToken.UrlTokenTokenType) {
+    // if (token.typ == EnumToken.UrlTokenTokenType) {
+    //
+    //     // @ts-ignore
+    //     return {
+    //         valid: SyntaxValidationResult.Valid,
+    //         context: [],
+    //         node: token,
+    //         // @ts-ignore
+    //         syntax: 'url()',
+    //         error: ''
+    //     }
+    // }
 
-        // @ts-ignore
-        return {
-            valid: SyntaxValidationResult.Valid,
-            context: [],
-            node: token,
-            // @ts-ignore
-            syntax: 'url()',
-            error: ''
-        }
-    }
-
-    if (token.typ != EnumToken.UrlFunctionTokenType) {
-
-        // @ts-ignore
-        return {
-            valid: SyntaxValidationResult.Drop,
-            context: [],
-            node: token,
-            // @ts-ignore
-            syntax: 'url()',
-            error: 'expected url()'
-        }
-    }
+    // if (token.typ != EnumToken.UrlFunctionTokenType) {
+    //
+    //     // @ts-ignore
+    //     return {
+    //         valid: SyntaxValidationResult.Drop,
+    //         context: [],
+    //         node: token,
+    //         // @ts-ignore
+    //         syntax: 'url()',
+    //         error: 'expected url()'
+    //     }
+    // }
 
     const children = (token as FunctionURLToken).chi.slice() as Token[];
 
     consumeWhitespace(children);
 
-    if (children.length == 0 || ![EnumToken.UrlTokenTokenType, EnumToken.StringTokenType, EnumToken.HashTokenType].includes(children[0].typ)) {
+    if (children.length  > 0 && [EnumToken.UrlTokenTokenType, EnumToken.StringTokenType, EnumToken.HashTokenType].includes(children[0].typ)) {
 
+        children.shift();
         // @ts-ignore
-        return {
-            valid: SyntaxValidationResult.Drop,
-            context: [],
-            node: children[0] ?? token,
-            // @ts-ignore
-            syntax: 'url()',
-            error: 'expected url-token'
-        }
+        // return {
+        //     valid: SyntaxValidationResult.Drop,
+        //     context: [],
+        //     node: children[0] ?? token,
+        //     // @ts-ignore
+        //     syntax: 'url()',
+        //     error: 'expected url-token'
+        // }
     }
-
-    children.shift();
 
     consumeWhitespace(children);