+++ src/lib/validation/utils/list.ts
@@ -8,10 +8,10 @@ export function stripCommaToken(tokenList: Token[]): Token[] | null {
 
     for (let i = 0; i < tokenList.length; i++) {
 
-        if (tokenList[i].typ == EnumToken.CommaTokenType && last != null && last.typ == EnumToken.CommaTokenType) {
-
-            return null;
-        }
+        // if (tokenList[i].typ == EnumToken.CommaTokenType && last != null && last.typ == EnumToken.CommaTokenType) {
+        //
+        //     return null;
+        // }
 
         if (tokenList[i].typ != EnumToken.WhitespaceTokenType) {
 
@@ -33,10 +33,10 @@ export function splitTokenList(tokenList: Token[], split: EnumToken[] = [EnumTok
 
     return tokenList.reduce((acc: Token[][], curr: Token): Token[][] => {
 
-        if (curr.typ == EnumToken.CommentTokenType) {
-
-            return acc;
-        }
+        // if (curr.typ == EnumToken.CommentTokenType) {
+        //
+        //     return acc;
+        // }
 
         if (split.includes(curr.typ)) {